applied code review suggestions
