+++ lib/src/lints/avoid_global_state/avoid_global_state_rule.dart
@@ -1,9 +1,8 @@
 import 'package:analyzer/analysis_rule/analysis_rule.dart';
 import 'package:analyzer/analysis_rule/rule_context.dart';
 import 'package:analyzer/analysis_rule/rule_visitor_registry.dart';
-import 'package:analyzer/dart/ast/ast.dart';
-import 'package:analyzer/dart/ast/visitor.dart';
 import 'package:analyzer/error/error.dart';
+import 'package:solid_lints/src/lints/avoid_global_state/avoid_global_state_rule_visitor.dart';
 
 /// Avoid top-level and static mutable variables.
 ///
@@ -51,8 +50,7 @@ class AvoidGlobalStateRule extends AnalysisRule {
   AvoidGlobalStateRule()
       : super(
           name: lintName,
-          description:
-              'Avoid top-level or static mutable variables to reduce shared mutable state.',
+          description: 'Avoid top-level or static mutable variables ',
         );
 
   @override
@@ -63,56 +61,9 @@ class AvoidGlobalStateRule extends AnalysisRule {
     RuleVisitorRegistry registry,
     RuleContext context,
   ) {
-    final visitor = _Visitor(this);
+    final visitor = AvoidGlobalStateRuleVisitor(this);
 
     registry.addTopLevelVariableDeclaration(this, visitor);
     registry.addFieldDeclaration(this, visitor);
-  }
-}
-
-class _Visitor extends SimpleAstVisitor<void> {
-  final AvoidGlobalStateRule rule;
-
-  _Visitor(this.rule);
-
-  @override
-  void visitTopLevelVariableDeclaration(TopLevelVariableDeclaration node) {
-    for (final variable in node.variables.variables) {
-      if (_isPublicMutable(variable)) {
-        rule.reportAtNode(variable);
-      }
-    }
-  }
-
-  @override
-  void visitFieldDeclaration(FieldDeclaration node) {
-    if (!node.isStatic) return;
-
-    for (final variable in node.fields.variables) {
-      if (_isPublicMutable(variable)) {
-        rule.reportAtNode(variable);
-      }
-    }
-  }
-
-  /// Returns true if the variable is mutable and not private.
-  bool _isPublicMutable(VariableDeclaration variable) {
-    return _isMutable(variable) && !_isPrivate(variable);
-  }
-
-  /// A variable is mutable if it is not final or const.
-  bool _isMutable(VariableDeclaration variable) {
-    final element = variable.declaredFragment?.element;
-
-    final isFinal = element?.isFinal ?? false;
-    final isConst = element?.isConst ?? false;
-
-    return !isFinal && !isConst;
-  }
-
-  /// A variable is private if its element is private.
-  bool _isPrivate(VariableDeclaration variable) {
-    final element = variable.declaredFragment?.element;
-    return element?.isPrivate ?? false;
   }
 }
