+++ lint_test/pubspec.yaml
@@ -12,6 +12,4 @@ dependencies:
 dev_dependencies:
   solid_lints:
     path: ../
-  analyzer_testing: ^0.1.9
-  test_reflective_loader: ^0.3.0
-  test: ^1.25.0
+
