+++ pubspec.yaml
@@ -24,4 +24,7 @@ dependencies:
 
 dev_dependencies:
   args: ^2.6.0
+  analyzer_testing: ^0.1.9
+  test_reflective_loader: ^0.3.0
+  test: ^1.25.0
