+++ test/avoid_global_state_rule_test.dart
@@ -2,12 +2,17 @@ import 'package:analyzer_testing/analysis_rule/analysis_rule.dart';
 import 'package:solid_lints/src/lints/avoid_global_state/avoid_global_state_rule.dart';
 import 'package:test_reflective_loader/test_reflective_loader.dart';
 
+void main() {
+  defineReflectiveSuite(() {
+    defineReflectiveTests(AvoidGlobalStateRuleTest);
+  });
+}
+
 @reflectiveTest
 class AvoidGlobalStateRuleTest extends AnalysisRuleTest {
   @override
   void setUp() {
     rule = AvoidGlobalStateRule();
-
     super.setUp();
   }
 
@@ -31,26 +36,41 @@ class Test {
     );
   }
 
-  void test_does_not_report_immutable_or_private_cases() async {
+  void test_does_not_report_global_immutable_variables() async {
     await assertNoDiagnostics(r'''
 final globalFinal = 1;
 const globalConst = 1;
+''');
+  }
 
+  void test_does_not_report_global_private_variables() async {
+    await assertNoDiagnostics(r'''
 var _privateTopLevel = 0;
+''');
+  }
 
+  void test_does_not_report_class_level_immutable_variables() async {
+    await assertNoDiagnostics(r'''
 class Test {
   static final int staticFinal = 1;
   static const int staticConst = 2;
-  static int _staticPrivate = 0;
-  final int instanceFinal = 1;
-  int instanceMutable = 0;
 }
 ''');
   }
-}
 
-void main() {
-  defineReflectiveSuite(() {
-    defineReflectiveTests(AvoidGlobalStateRuleTest);
-  });
+  void test_does_not_report_class_level_private_variables() async {
+    await assertNoDiagnostics(r'''
+class Test {
+  static int _staticPrivate = 0;
+}
+''');
+  }
+
+  void test_does_not_report_local_method_variables() async {
+    await assertNoDiagnostics(r'''
+void m() {
+  int localMutable = 0;
+}
+''');
+  }
 }
